version 2.2.6, close #137
- enhance identify overlay package
- fix auto-brevent safe package in non-aggressive mode

+++ brevent/src/main/res/raw/brevent.sh
@@ -1,5 +1,5 @@
 # path, abi64 should have been defined
-if [ ! -f $path -o -z $abi64 ]; then
+if [ -z $abi64 ]; then
     echo "ERROR: please open Brevent to make a new brevent.sh" >&2
     exit 1
 fi
@@ -7,18 +7,23 @@ fi
 brevent=/data/local/tmp/brevent
 
 # some os cannot execute $path directly
-if [ -f $brevent ]; then
-    rm -rf $brevent
-fi
-if [ -f $brevent ]; then
-    echo "WARNING: /data/local/tmp is not writable" >&2
-else
-    cp $path $brevent
+if [ -x $path ]; then
     if [ -f $brevent ]; then
-        chmod 0755 $brevent
-    else
-        echo "WARNING: /data/local/tmp is not writable" >&2
+        rm -rf $brevent
     fi
+    if [ -f $brevent ]; then
+        echo "WARNING: /data/local/tmp is not writable" >&2
+    else
+        cp $path $brevent
+        if [ -f $brevent ]; then
+            chmod 0755 $brevent
+        else
+            echo "WARNING: /data/local/tmp is not writable" >&2
+        fi
+    fi
+elif [ ! -x $brevent ]; then
+    echo "ERROR: please open Brevent to make a new brevent.sh" >&2
+    exit 1
 fi
 
 # some os is 64bit, but load 32bit library(and binary)
